backend/requirements.txt
flask
flask_cors
mysql-connector-python
bcrypt
python-dotenv